# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ScopeWise AI API",
        "description": "REST API for software requirement analysis, scope-creep prevention and project estimation. Authenticate with a Sanctum bearer token obtained from POST /api/auth/login.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{scheme}://{host}",
            "description": "Current environment",
            "variables": {
                "scheme": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8000"
                }
            }
        }
    ],
    "paths": {
...
        "/api/analysis/{id}/questions": {
            "get": {
                "tags": [
                    "Analysis"
                ],
                "summary": "Get the clarification questions for an analysis",
                "operationId": "db2881dd603920715ea534ff5be61f2f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of questions with stable ids (q-1 … q-n)"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not your analysis"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/analysis/{id}/questions/{questionId}": {
            "post": {
                "tags": [
                    "Analysis"
                ],
                "summary": "Answer a clarification question",
                "description": "Recording the answer recomputes the analysis confidence.",
                "operationId": "aef516de29d54af29bfd273a1de73095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "questionId",
                        "in": "path",
                        "description": "e.g. q-1",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "answer"
                                ],
                                "properties": {
                                    "answer": {
                                        "type": "string",
                                        "example": "Yes, admins can also expire sessions"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated analysis"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not your analysis"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Sanctum token issued by POST /api/auth/login. Send the token as: Authorization: Bearer <token>"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}